requirements.txt
paho-mqtt==1.5.1
pyserial==3.5
rpi-rf==0.9.7
RPi.GPIO==0.7.0
